requirement.txt
pinecone>=5.0.0
sentence-transformers>=3.0.0
pandas>=2.1.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
